Crate oak_https_attestation

HTTPS Attestation Service implementation.

HTTPS Attestation Service works as an HTTPS proxy for a TEE application that runs inside the same TEE as the application.

HTTPS Attestation Service allows to perform bidirectional remote attestation via TLS by cryptographically binding assertions to a TLS session, using Exported Keying Material from an already established TLS session and producing an assertion based on it, e.g. signing it with a TEE key.

HTTPS Attestation Service description is provided in the TLS Remote Attestation RFC.

Note: Current version of the HTTPS Attestation Service doesn’t have remote attestation and client assertion generation.

Modules

server

Structs

Opt

Functions

load_certificate

load_private_key

main